I know, I'm doing it wrong...
Add Authentication System
* AuthenticationModel: scaffold class for the user model
* AuthenticationController: scaffold class for the user controller
* AuthenticatedController: integrates authentication into the Controller class
* views/authentication: scaffold views
* helpers/authentication_helper.php: is_logged_in() and is_admin() helpers
* config.php: add 'auth_model' and 'auth_controller' settings

Controller
* get scaffold model name from controller name
* only use public instance methods as actions
* add has_action() to check if an action exists or is provided by scaffolding
* perform():
  * only check for valid characters in action
  * add before-render filters
  * cache(): let View decide if caching is enabled
* scaffold():
  * don't redirect to non-existent actions
  * split actual action implementations into seperate methods

Dispatcher:
* don't collect render time
* log footer before adding debug toolbar

Model
* __toString(): return a text representation of the model, replaces get_display_name()
* add $_display_column, to control __toString()
* add to_param(), to_json(), to_xml()
* add basic auto_field()
* combine changed() and get_changed()
* remove cache stuff
* add generic ModelMapper
* merge stuff from ActiveRecord:
  * virtual attributes
  * new_record() / exists()
  * load() and reload()
  * save() and destroy()
* add link_to() and button_to()

Database
* DatabaseConnection
  * add inspect()
  * improve parse_type()
* ActiveRecord
  * add get_slug() to add ID to slug
  * remove stuff merged into Model
* DatabaseMapper
  * remove stuff merged into ModelMapper
  * add find(), find_all()
  * add delete_all(), remove destroy()
  * build_conditions(): allow object values
* QuerySet
  * add inspect(), to_json(), to_xml()
  * add destroy_all()
  * add without(), without_blobs()
  * support array values for SELECT and FROM
  * only use first value for LIMIT and OFFSET
  * add has_key() to check for table attributes and aliased keys
  * sorted(), filtered(): use has_key()

View
* add View::$current to keep track of the current view instance, assign to 'view'
* View::find_template(): allow absolute paths
* set(): allow omitting key name for object values, use tableized class or model name
* add add() to set multiple default values
* cache(): check if view caching is enabled
* add log() to clean up log messages
* compile(): use a handler if available

Inflector
* humanize(): automatically translate text
* pluralize(): use ngettext()
* use mb_* functions

Debug Toolbar
* add homepage link to menu
* _request.thtml: show authentication information
* _template.thtml: hide uninteresting values, use dump_value()

Debug Controller
* use tool icons
* Model Browser:
  * localize messages
  * don't add abstract models

Configuration
* add documentation
* always show errors in console, use exception handler
* only attach a fatal error handler if an exception handler is set
* set_language(): allow language 'C' to reset locale
* use $_ENV['ENVIRONMENT'] if available
* __autoload(): try loading models from LIB

Errors
* ValueError, TypeError: allow custom message
* log NotFound errors with LOG_INFO

Helpers
* use automatic pluralize()
* text_helper.php: add to_string(), to_json() and to_xml()
* debug_helper.php: add dump_value(), dump_function()
* asset_helper.php: add icon(), allow relative paths in asset_path()
* add new Textile class
* truncate(): escape values if $add_title is true
* list_tag(), table_tag(): escape values by default
* url_for(): allow passing relative paths and objects
* link_to(): add JS form token if necessary
* check for controller instance before using Dispatcher::$controller

Router
* generate(): add object values, using to_params() and to_param()

Testing
* test/runner.php: recursively load test helpers, load fixtures correctly, only run in test environment
* test/coverage.php: use property prefixes

Scripts
* move fake_request() to lib/script.php
* add script/doc, a no-bullshit documentation generator
* script/translate: add model names
* script/console: remove stuff merged into helpers

Utilities
* add util/init.php to ensure correct loading order
* util/objects.php:
  * replace __toString() with inspect(), add simple __toString()
  * __get(): allow using a method name directly
* util/system.php: add check_root() to compare two paths, and is_command() to check if a command exists
* util/arrays.php: add first(), last(), sorted()
* add util/hash.php

+++ lib/database/mapper.php
@@ -10,7 +10,7 @@
    require LIB.'database/query_set.php';
    require LIB.'database/association.php';
 
-   abstract class DatabaseMapper extends Object
+   abstract class DatabaseMapper extends ModelMapper
    {
       static function load($model) {
          static $_cache;
@@ -25,14 +25,11 @@
          }
       }
 
-      protected $_model;
       protected $_table;
       protected $_primary_key = 'id';
       protected $_database = 'default';
       protected $_connection;
 
-      protected $_attributes;
-      protected $_defaults;
       protected $_order;
       protected $_page_size;
 
@@ -42,11 +39,11 @@
       protected $_belongs_to;
 
       function __construct() {
-         $this->_model = substr(get_class($this), 0, -6);
+         parent::__construct();
 
          if (empty($this->_database)) {
             throw new ConfigurationError("No database set for model '{$this->_model}'");
-         } elseif (empty($this->_table)) {
+         } elseif (empty($this->_table) and !($this->_table = tableize($this->_model))) {
             throw new ConfigurationError("No table set for model '{$this->_model}'");
          }
 
@@ -61,16 +58,12 @@
          }
       }
 
-      function __toString() {
-         return parent::__toString(array(
+      function inspect() {
+         return parent::inspect(array(
             'model'    => $this->_model,
             'database' => $this->_database,
             'table'    => $this->_table,
          ));
-      }
-
-      function get_model() {
-         return $this->_model;
       }
 
       function get_database() {
@@ -101,10 +94,6 @@
          }
       }
 
-      function get_defaults() {
-         return $this->_defaults;
-      }
-
       function get_order() {
          return $this->_order;
       }
@@ -127,12 +116,41 @@
 
       function execute() {
          $args = func_get_args();
-         return call_user_func_array(array($this->connection, execute), $args);
+         return call_user_func_array(array($this->connection, 'execute'), $args);
       }
 
-      function create(array $attributes) {
-         $model = new $this->_model($attributes);
-         return $model->save();
+      function get_query_set() {
+         if (class_exists($class = $this->_model.'QuerySet')) {
+            return new $class($this);
+         } else {
+            return new QuerySet($this);
+         }
+      }
+
+      # Forward accesses to non-existent properties to a QuerySet
+      function __get($key) {
+         if (method_exists($this, $setter = "get_$key")) {
+            return $this->$setter();
+         } elseif (method_exists($this, $key)) {
+            return $this->$key();
+         } else {
+            return $this->get_query_set()->$key;
+         }
+      }
+
+      # Forward calls to non-existent methods to a QuerySet
+      function __call($method, $args) {
+         return call_user_func_array(array($this->get_query_set(), $method), $args);
+      }
+
+      function find($conditions) {
+         $args = func_get_args();
+         return $this->__call('find', $args);
+      }
+
+      function find_all($conditions=null) {
+         $args = func_get_args();
+         return $this->__call('where', $args);
       }
 
       function insert(array $attributes) {
@@ -159,11 +177,13 @@
                           $this->_table, implode(", ", $columns), implode(", ", $keys));
 
          if ($this->execute($query, $values)) {
-            return $this->connection->insert_id();
+            return array(
+               $this->_primary_key => $this->connection->insert_id()
+            );
          }
       }
 
-      function update(array $attributes, $conditions, $force=false) {
+      function update($conditions, array $attributes, $force=false) {
          if (empty($attributes) and !$force) {
             return true;
          }
@@ -186,7 +206,10 @@
             return true;
          }
 
-         list($conditions, $condition_values) = $this->build_condition($conditions);
+         list($conditions, $condition_values) = $this->build_condition(
+            is_object($conditions) ? $conditions->id : $conditions
+         );
+
          if (blank($conditions)) {
             throw new ApplicationError("No conditions given");
          } else {
@@ -199,7 +222,7 @@
          return is_object($this->execute($query, $values));
       }
 
-      function delete($id) {
+      function delete($conditions) {
          list($conditions, $values) = $this->build_condition(func_get_args());
          if (blank($conditions)) {
             throw new ApplicationError("No conditions given");
@@ -208,62 +231,18 @@
          return is_object($this->execute("DELETE FROM `{$this->_table}` WHERE $conditions", (array) $values));
       }
 
-      function destroy($conditions) {
-         $args = func_get_args();
-         $objects = call_user_func_array(
-            array($this->get_query_set(), 'where'), $args
-         );
-
-         $status = false;
-         foreach ($objects as $object) {
-            $status = $object->destroy();
-         }
-
-         return $status;
-      }
-
-      function delete_all() {
-         $this->execute("DELETE FROM `{$this->_table}`");
-      }
-
-      function get_query_set() {
-         if (class_exists($class = $this->_model.'QuerySet')) {
-            return new $class($this);
+      function delete_all($force=false) {
+         if ($force) {
+            $this->execute("DELETE FROM `{$this->_table}`");
          } else {
-            return new QuerySet($this);
+            throw new ApplicationError("Not unless you really want to");
          }
-      }
-
-      function __get($key) {
-         $getter = "get_$key";
-         if (method_exists($this, $getter)) {
-            return $this->$getter();
-         } else {
-            return $this->get_query_set()->$getter();
-         }
-      }
-
-      function __set($key, $value) {
-         $setter = "set_$key";
-         if (method_exists($this, $setter)) {
-            $this->$setter($value);
-            return $this;
-         } else {
-            return $this->get_query_set()->$setter($key);
-         }
-      }
-
-      function __call($method, $args) {
-         return call_user_func_array(array($this->get_query_set(), $method), $args);
       }
 
       function build_condition($conditions) {
          if (!is_array($conditions)) {
             # Conditions are passed directly as arguments
             $conditions = func_get_args();
-         } elseif (count($conditions) == 1 and is_array($conditions[0])) {
-            # Conditions are passed as array
-            $conditions = $conditions[0];
          }
 
          if (empty($conditions)) {
@@ -279,7 +258,7 @@
 
          while (!empty($keys)) {
             $key = array_shift($keys);
-            $value = array_shift($values);
+            $value = $this->convert(array_shift($values));
 
             if ($condition and $operator == '') {
                $operator = ' AND ';
@@ -306,12 +285,12 @@
                   $params[] = $value;
                }
 
-            } elseif (is_numeric($value)) {
+            } elseif (is_numeric($value) or is_numeric($value[0])) {
                # Use array value as ID
                #   e.g.: find($id)
                #
                $condition .= "$operator`{$this->_table}`.`{$this->_primary_key}` = ?";
-               $params[] = $value;
+               $params[] = intval($value);
 
                # Allow passing multiple IDs
                $operator = ' OR ';
@@ -322,7 +301,7 @@
                #
                $condition .= "$operator$value";
                for ($i = 0; $i < $count; $i++) {
-                  $params[] = array_shift_arg($values);
+                  $params[] = $this->convert(array_shift_arg($values));
                   array_shift($keys);
                }
 
@@ -339,7 +318,7 @@
                # Use array value as column name
                #   e.g.: find('key', $value)
                $condition .= "$operator$value = ?";
-               $params[] = array_shift_arg($values);
+               $params[] = $this->convert(array_shift_arg($values));
                array_shift($keys);
 
             } elseif (!is_null($value)) {
@@ -348,6 +327,20 @@
          }
 
          return array($condition, $params);
+      }
+
+      protected function convert($value) {
+         if (is_object($value)) {
+            if (!$value instanceof ActiveRecord) {
+               throw new TypeError($value, "Invalid class '%s'");
+            } elseif ($value->changed) {
+               throw new ApplicationError("Can't use unsaved model as parameter");
+            } else {
+               return $value->id;
+            }
+         } else {
+            return $value;
+         }
       }
    }
